Created new folder and moved files inside
