+++ dotcom-rendering/src/web/components/Callout/FormFields/FileUpload.tsx
@@ -1,8 +1,8 @@
 import { css } from '@emotion/react';
 import { space, text, textSans } from '@guardian/source-foundations';
 import { useState } from 'react';
-import { CampaignFieldFile } from '../../../types/content';
-import { stringifyFileBase64 } from '../../lib/stringifyFileBase64';
+import { CampaignFieldFile } from '../../../../types/content';
+import { stringifyFileBase64 } from '../../../lib/stringifyFileBase64';
 import { FieldLabel } from './FieldLabel';
 
 const fileUploadInputStyles = css`
