+++ dotcom-rendering/src/web/components/Callout/FormFields/MultiSelect.tsx
@@ -1,7 +1,7 @@
 import type {
 	CampaignFieldCheckbox,
 	CampaignFieldRadio,
-} from '../../../types/content';
+} from '../../../../types/content';
 import { CheckboxSelect } from './CheckboxSelect';
 import { RadioSelect } from './RadioSelect';
